+++ modules/webtools/functions/@weboptions/weboptions.m
@@ -48,6 +48,7 @@ function options = buildDefaults()
   else
     options.CertificateFilename = [];
   end
+  options.FollowLocation = false;
 end
 %=============================================================================
 function res = isSupportedFieldNames(argName)
@@ -64,7 +65,8 @@ function res = isSupportedFieldNames(argName)
   'MediaType';
   'RequestMethod';
   'Arrayformat';
-  'CertificateFilename'};
+  'CertificateFilename';
+  'FollowLocation'};
   res = any(contains(upper(supportedOptions), upper(argName)));
 end
 %=============================================================================
@@ -101,6 +103,8 @@ function [res, fieldname, fieldvalue] = checkArgument(argName, argValue)
       [res, fieldname, fieldvalue] = checkHeaderFields(argValue);
     case 'CERTIFICATEFILENAME'
       [res, fieldname, fieldvalue] = checkCertificateFilename(argValue);
+    case 'FOLLOWLOCATION'
+      [res, fieldname, fieldvalue] = checkFollowLocation(argValue);
     otherwise
       res = false;
       fieldname = argName;
@@ -253,5 +257,11 @@ function [res, nameNormalized, valueNormalized] = checkCertificateFilename(argVa
       res = isempty(valueNormalized) || isfile(valueNormalized);
     end
   end
+end
+%=============================================================================
+function [res, nameNormalized, valueNormalized] = checkFollowLocation(argValue)
+  res = islogical(argValue) && isscalar(argValue) && ~issparse(argValue);
+  valueNormalized = argValue;
+  nameNormalized = 'FollowLocation';
 end
 %=============================================================================