+++ modules/webtools/tests/test_webread_7.m
@@ -7,21 +7,16 @@
 % SPDX-License-Identifier: LGPL-3.0-or-later
 % LICENCE_BLOCK_END
 %=============================================================================
-function names = fieldnames(options)
-  names = {'CharacterEncoding';
-  'UserAgent';
-  'Timeout';
-  'Username';
-  'Password';
-  'KeyName';
-  'KeyValue';
-  'HeaderFields';
-  'ContentType';
-  'ContentReader';
-  'MediaType';
-  'RequestMethod';
-  'Arrayformat';
-  'CertificateFilename';
-  'FollowLocation'};
+if ismodule('audio')
+  o = weboptions('ContentType', 'binary', 'ContentReader', str2func('audioread'), 'FollowLocation', true);
+  o.Timeout = 120;
+  try
+    R = webread('https://freewavesamples.com/files/Ensoniq-ZR-76-01-Dope-77.wav', o);
+  catch ex
+    R = strcmp(ex.message, _('Forbidden (403)')) || ...
+        strcmp(ex.message, _('Timeout was reached')) || ... 
+        strcmp(ex.message, _('Couldn''t resolve host name'));
+    skip_testsuite(R, ex.message)
+  end
 end
 %=============================================================================