+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `nthroot`: Real nth root of real numbers.
 - `allfinite`: Check if all array elements are finite.
 - `j` as imaginary unit number is also supported. example `3+2j` equivalent to `3+2i`.
+- `FollowLocation` option for `weboptions`
 - oneAPI Threading Building Blocks optional dependency.
 
 ### Changed